creates header for diff_drive_controller
